[nanobc] Resolves issue #33.
When `SQLExecute()` returns `SQL_NO_DATA`, subsequent calls to functions such as `SQLNumResultCols()` or `SQLRowCount()` can result in a driver error. For example, this occurs on vertica drivers.

As such, I have added a CMake option `NANODBC_HANDLE_NODATA_BUG` that, when enabled, compiles for support for drives such as the vertica driver.